+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `LockBundle`, which provides the `lock` configuration and the services previously provided by `FrameworkBundle` under `framework.lock`
+ * Pick the default lock store when the bundle loads instead of in the config tree; an empty `resources` list now means the default store
  * Add argument `$advisory` to `StoreFactory::createStore()` to use advisory locks when reusing an existing `\PDO` or Doctrine DBAL `Connection`
  * Add `DoctrineDbalMysqlStore` based on MySQL `GET_LOCK()` functionality, usable with the `mysql+advisory://` DSN
  * Add `MysqlStore` based on MySQL `GET_LOCK()` functionality, usable with the `mysql+advisory:` DSN